backend/python-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
pydantic>=2.7,<3
python-dotenv==1.0.1
google-generativeai>=0.8.0
pdf2image==1.17.0
openpyxl==3.1.5
SQLAlchemy==2.0.23
pydantic-settings==2.4.0
openai>=1.81.0,<2
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn>=1.3
PyPDF2>=3.0.0
pdfminer.six>=20231228
pytesseract>=0.3.10
Pillow>=9.5
